[shop] revert 스테이지 생성 및 확정 이벤트 컨슈머 생성

+++ src/main/kotlin/gogo/gogoshop/domain/cointoss/application/dto/CoinTossDto.kt
@@ -2,6 +2,6 @@ package gogo.gogoshop.domain.cointoss.application.dto
 
 data class CoinTossResDto(
     val coinTossId: Long,
-    val ticketPrice: Int,
+    val ticketPrice: Long,
     val ticketQuantity: Int
 )
