+++ src/main/kotlin/gogo/gogoshop/domain/yavarwee/application/YavarweeReader.kt
@@ -17,7 +17,7 @@ class YavarweeReader(
         return yavarwee!!.yavarweeId
     }
 
-    fun readYavarweeTicketPrice(shopId: Long): Int {
+    fun readYavarweeTicketPrice(shopId: Long): Long {
         val yavarwee = read(shopId)
         return yavarwee!!.ticketPrice
     }
